pyproject.toml
[project]
name = "zendesk-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "hydra-core>=1.3.2",
    "jupyter>=1.1.1",
    "pandas>=2.3.3",
    "requests>=2.32.5",
]
